+++ capitalize_decorator.rb
@@ -1,0 +1,5 @@
+class CapitalizeDecorator < BaseDecorator
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
